delete user security

+++ admin/components/view_all_users.php
@@ -78,12 +78,17 @@
 
 	// Deleting Users
 	if(isset($_GET['delete'])) {
+		if(isset($_SESSION['user_role'])) {
+			if($_SESSION['user_role'] == 'admin') {
+				$the_user_id = mysqli_real_escape_string($sonnection, $_GET['delete']);
 		
-		$the_user_id = $_GET['delete'];
+				$query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
+				$delete_user_query = mysqli_query($connection, $query);
+				
+				header("location: users.php");
+			}
+		}
 		
-		$query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
-		$delete_user_query = mysqli_query($connection, $query);
-		
-		header("location: users.php");
+
 	}
 ?>
